requirements.txt
numpy~=1.20.3
Pillow~=8.2.0
PyPrind~=2.11.3
pygame~=2.0.1